+++ app/controllers/concerns/edge_validator.rb
@@ -46,4 +46,9 @@ module EdgeValidator
     end
   end
 
+  private
+    def numeric? val
+      val == val.to_i.to_s
+    end
+
 end