+++ app/controllers/edge/internal/edge_authenticators_controller.rb
@@ -7,8 +7,6 @@ class EdgeAuthenticatorsController < RestController
   include BodyParser
   include Cryptography
   include EdgeValidator
-  include ExtractEdgeResources
-  include GroupMembershipValidator
   include AuthenticatorsManager
   def all_authenticators
     logger.info(LogMessages::Endpoints::EndpointRequested.new("all-authenticators"))
@@ -19,21 +17,24 @@ class EdgeAuthenticatorsController < RestController
       verify_edge_host(options)
       verify_kind(options[:kind])
       kinds = options[:kind].split(',')
-      scope = get_authenticators_data(kinds)
       if params[:count] == 'true'
-        count=params[:count]
+        scope = get_authenticators_data(kinds)
         count_authenticators={}
         scope.each do |authenticator_kind, authenticators_list|
           count_authenticators[authenticator_kind] = authenticators_list.size
         end
         results = { count: count_authenticators }
-        logger.info(LogMessages::Endpoints::EndpointFinishedSuccessfullyWithCount.new("all-authenticators", count))
+        logger.info(LogMessages::Endpoints::EndpointFinishedSuccessfullyWithCount.new("all-authenticators", params[:count]))
         render(json: results)
       else
+        offset = options[:offset]
+        limit = options[:limit]
+        validate_scope(limit, offset)
         verify_header(request)
         response.set_header("Content-Encoding", "base64")
+        parsed_data = get_authenticators_parsed_data(kinds, offset, limit)
         logger.info(LogMessages::Endpoints::EndpointFinishedSuccessfully.new("all-authenticators"))
-        render(json: scope)
+        render(json: parsed_data)
       end
     rescue ApplicationController::Forbidden
       raise
@@ -47,8 +48,9 @@ class EdgeAuthenticatorsController < RestController
   private
   def verify_kind(kinds_param)
     allowed_kind = ['authn-jwt']
-    kinds = kinds_param.split(',')
-    unless kinds.all? { |value| allowed_kind.include?(value) }
+    kinds = kinds_param.to_s.split(',')
+    # the kind param cant be empty, and the values have to be from the allowed_kind list
+    unless kinds.present? && kinds.all? { |value| allowed_kind.include?(value) }
       raise ArgumentError , "authenticator kind parameter is not valid"
     end
   end