+++ app/controllers/secrets_controller.rb
@@ -184,7 +184,7 @@ class SecretsController < RestController
     # There shouldn't be a state where a variable belongs to an issuer that doesn't exit, but we check it to be safe
     raise ApplicationController::InternalServerError, "Issuer assigned to #{account}:#{params[:kind]}:#{params[:identifier]} was not found" unless issuer
 
-    logger.info(LogMessages::Secrets::EphemeralSecretRequest.new(variable_data["issuer"], issuer.issuer_type, variable_data["method"], request_id))
+    logger.info(LogMessages::Secrets::EphemeralSecretRequest.new(request_id, variable_data["issuer"], issuer.issuer_type, variable_data["method"]))
 
     issuer_data = {
       max_ttl: issuer.max_ttl,