+++ app/domain/edge_logic/data_handlers/ongoing_handler.rb
@@ -25,7 +25,7 @@ module EdgeLogic
         #convert time to seconds 
         last_synch_time_sec = Rational(stats['last_synch_time'], 1000)
         installation_time_sec = Rational(edge.installation_date, 1000)
-        @logger.info(LogMessages::Edge::EdgeTelemetry.new(parsed_tenant_id, edge.name, Time.at(last_synch_time_sec),
+        @logger.info(LogMessages::Edge::EdgeTelemetry.new(tenant_id, edge.name, Time.at(last_synch_time_sec),
                                                          cycle_reqs['get_secret'], cycle_reqs['apikey_authenticate'],
                                                          cycle_reqs['jwt_authenticate'], cycle_reqs['redirect'],
                                                          edge.version, edge.platform, Time.at(installation_time_sec)))
@@ -48,14 +48,8 @@ module EdgeLogic
 
       private
 
-      def parsed_tenant_id
-        tenant_id = Rails.application.config.conjur_config.tenant_id
-        if tenant_id.match?(/\A[a-f0-9]{32}\z/)
-          # Insert dashes at the specific positions to achieve the tenant-id format
-          formatted_output = "#{tenant_id[0..7]}-#{tenant_id[8..11]}-#{tenant_id[12..15]}-#{tenant_id[16..19]}-#{tenant_id[20..31]}"
-        else
-          ""
-        end
+      def tenant_id
+        Rails.application.config.conjur_config.tenant_id
       end
 
     end