+++ config/routes.rb
@@ -67,7 +67,7 @@ Rails.application.routes.draw do
       get     "/resources/:account"                   => "resources#index"
       get     "/resources"                            => "resources#index"
 
-      post    "hosts/:account/*identifier"            => "hosts#post"
+      post    "hosts/:account/*identifier"            => "workload#post"
 
       get     "/:authenticator/:account/providers"  => "providers#index"
 
@@ -90,7 +90,7 @@ Rails.application.routes.draw do
 
       post    "/edge/data/:account"                 => 'edge_handler#report_edge_data', :constraints => QueryParameterActionRecognizer.new("data_type")
       get     "/edge/slosilo_keys/:account"         => 'edge_slosilo_keys#slosilo_keys'
-
+      get     "/features"                       => "feature_flag#feature_flag"
       post    "/issuers/:account"             => 'issuers#create'
       delete  "/issuers/:account/:identifier" => 'issuers#delete'
       get     "/issuers/:account/:identifier" => 'issuers#get'