+++ cucumber/api/features/step_definitions/audit_steps.rb
@@ -18,6 +18,14 @@ Then(/^there is an audit record matching:$/) do |given|
   end
 end
 
+Then(/^there is no audit record matching:$/) do |given|
+  if Utils.local_conjur_server
+    expect(audit_messages).to !include(matching(audit_template(given)))
+  else
+    expect(num_matches_since_savepoint(normalized_to_log(given))).to be < 1
+  end
+end
+
 module CucumberAuditHelper
   def audit_messages
     Test::AuditSink.messages.map(&method(:normalized_message))