+++ CHANGELOG.md
@@ -9,12 +9,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Nothing should go in this section, please add to the latest unreleased version
   (and update the corresponding date), or add a new version.
 
-## [1.0.6-cloud] - 2023-09-04
+## [1.0.9-cloud] - 2023-10-08
+### Added
+- Add feature flag endpoint
+
+## [1.0.8-cloud] - 2023-10-01
+### Changed
+- New env variables
+
+## [1.0.7-cloud] - 2023-09-04
 ### Changed
 - Renamed `platforms` to `issuers`, changed internal structure of ephemeral secret requests and removed default issuer secret
   [ONYX-42993](https://ca-il-jira.il.cyber-ark.com:8443/browse/ONYX-42993)
-- Get all authenticators endpoint, will be used by edge for replication.
-- Modify edge logs.
+- Get all authenticators endpoint, will be used by edge for replication
+- Modify edge logs
+- Add limit and offset to get all authenticators endpoint [ONYX-44074](https://ca-il-jira.il.cyber-ark.com:8443/browse/ONYX-44074)
 
 ## [1.0.6-cloud] - 2023-08-21
 ### Changed